make apply patches its own method
